refactor: faq-schema type added
